Move add class form to classes-form component
