+++ Organizely/src/app/classes-page/classes-page.component.html
@@ -1,70 +1,1 @@
-<form [formGroup]="addCourseForm" (ngSubmit)="submit()">
-  <label for="courseName">Course Name</label>
-  <input type="text" id="courseName" formControlName="courseName" />
-
-  <br />
-
-  <label for="startTime">Start Time</label>
-  <input type="time" id="startTime" formControlName="startTime" />
-
-  <br />
-
-  <label for="endTime">End Time</label>
-  <input type="time" id="endTime" formControlName="endTime" />
-
-  <br />
-
-  <label for="startRecur">Start Date</label>
-  <input type="date" id="startRecur" formControlName="startRecur" />
-
-  <br />
-
-  <label for="endRecur">End Date</label>
-  <input type="date" id="endRecur" formControlName="endRecur" />
-
-  <br />
-
-  <label>Days</label>
-
-  <br />
-
-  <label *ngFor="let day of dayNames">
-    {{ day["name"] }}
-    <input
-      type="checkbox"
-      [value]="day['id']"
-      (change)="onCheckChange($event)"
-    />
-    <br />
-  </label>
-
-  <br />
-
-  <label>Semester Season</label>
-
-  <select formControlName="semesterSeason">
-    <option *ngFor="let season of semesterSeasons" [value]="season">
-      {{ season }}
-    </option>
-  </select>
-
-  <br />
-
-  <label for="semesterYear">Semester Year</label>
-
-  <input
-    type="number"
-    id="semesterYear"
-    formControlName="semesterYear"
-    [min]="currentYear"
-  />
-
-  <br />
-
-  <label for="teacherName">Teacher's Name</label>
-  <input type="text" id="teacherName" formControlName="teacherName" />
-
-  <br />
-
-  <button type="submit">Submit</button>
-</form>
+<button routerLink="/organizely/classform">Add Class</button>
